+++ src/app.js
@@ -1,29 +1,62 @@
-import { Player } from "./objects/player.js";
-/** state of game is processing */
-let gameState = false;
-//#region game objects
-let player = null;
-let meteos = [];
-let planets = [];
-let items = [];
-let bullets = [];
-//#endregion
+import { Game } from "./game.js";
+var renderer;
+var USE_WIREFRAME = false;
 
-/** other objects like lights */
-let otherObjets = [];
-let renderer = null;
-let screen = {
-    scene: null,
-    camera: null,
+var loadingScreen = {
+    scene: new THREE.Scene(),
+    camera: new THREE.PerspectiveCamera(90, 1280 / 720, 0.1, 100),
+    box: new THREE.Mesh(new THREE.BoxGeometry(0.5, 0.5, 0.5), new THREE.MeshBasicMaterial({ color: 0x4444ff })),
 };
 
-/** keyboad pressed state */
-let keyboard = {};
+// Bullets array
 
 function init() {
-    window.addEventListener("keydown", keyDown);
-    window.addEventListener("keyup", keyUp);
-    resetGame();
+    window.game = new Game();
+    const game = window.game;
+
+    game.reset();
+
+    const light3 = new THREE.PointLight(0xc4c4c4, 0.8);
+    light3.position.set(50, 50, 50);
+
+    const scene = game.scene;
+    scene.add(light3);
+    //#region loading Screen
+    loadingScreen.box.position.set(0, 0, 5);
+    loadingScreen.camera.lookAt(loadingScreen.box.position);
+    loadingScreen.scene.add(loadingScreen.box);
+
+    //#endregion
+    const mesh = new THREE.Mesh(new THREE.BoxGeometry(1, 1, 1), new THREE.MeshPhongMaterial({ color: 0xff4444, wireframe: USE_WIREFRAME }));
+    mesh.position.y += 1;
+    mesh.receiveShadow = true;
+    mesh.castShadow = true;
+    scene.add(mesh);
+
+    const meshFloor = new THREE.Mesh(new THREE.PlaneGeometry(20, 20, 10, 10), new THREE.MeshPhongMaterial({ color: 0xffffff, wireframe: USE_WIREFRAME }));
+    meshFloor.rotation.x -= Math.PI / 2;
+    meshFloor.receiveShadow = true;
+    scene.add(meshFloor);
+
+    const ambientLight = new THREE.AmbientLight(0xffffff, 0.2);
+    scene.add(ambientLight);
+
+    const light = new THREE.PointLight(0xffffff, 0.8, 18);
+    light.position.set(-3, 6, -3);
+    light.castShadow = true;
+    light.shadow.camera.near = 0.1;
+    light.shadow.camera.far = 25;
+    scene.add(light);
+
+    // Load models
+    // REMEMBER: Loading in Javascript is asynchronous, so you need
+    // to wrap the code in a function and pass it the index. If you
+    // don't, then the index '_key' can change while the model is being
+    // downloaded, and so the wrong model will be matched with the wrong
+    // index key.
+    const loader = new THREE.TextureLoader();
+    const bgTexture = loader.load("res/bg.webp");
+    scene.background = bgTexture;
 
     renderer = new THREE.WebGLRenderer();
     renderer.setSize(1280, 720);
@@ -32,78 +65,94 @@ function init() {
     renderer.shadowMap.type = THREE.BasicShadowMap;
 
     document.body.appendChild(renderer.domElement);
-    const loader = new THREE.TextureLoader();
-    const bgTexture = loader.load("res/bg.webp");
-    screen.scene.background = bgTexture;
-    animate();
+
+    render();
 }
+// Runs when all resources are loaded
 
-/** reset game states */
-function resetGame() {
-    gameState = true;
-    // player = new Player(0,0,gltf, ()=>{});
-    screen.scene = new THREE.Scene();
-    screen.camera = new THREE.PerspectiveCamera(90, 1280 / 720, 0.1, 1000);
+function render() {
+    const game = window.game;
+    // Play the loading screen until resources are loaded.
+    if (!game.RESOURCES_LOADED) {
+        requestAnimationFrame(render);
+        loadingScreen.box.position.x -= 0.05;
+        if (loadingScreen.box.position.x < -10) loadingScreen.box.position.x = 10;
+        loadingScreen.box.position.y = Math.sin(loadingScreen.box.position.x);
+        renderer.render(loadingScreen.scene, loadingScreen.camera);
+        game.updateLoaded();
+        return;
+    }
 
-    // screen.scene.add(player.gltf.scene);
-    for (const meteo of meteos) screen.scene.add(meteo.gltf.scene);
-    for (const obj of otherObjets) screen.scene.add(obj.gltf.scene);
-    for (const planet of planets) screen.scene.add(planet.gltf.scene);
-    for (const item of items) screen.scene.add(item.gltf.scene);
-    for (const bullet of bullets) screen.scene.add(bullet.gltf.scene);
-}
+    const player = game.player;
+    player.update();
 
-function animate() {
-    generateObjects();
+    for (const meteor of game.meteors) meteor.move(game.scene, game.player);
+    for (const item of game.items) item.move(game.player);
+    for (const planet of game.planets) planet.move();
+    // Uncomment for absurdity!
+
+    // go through bullets array and update position
+    // remove bullets when appropriate
+    const bullets = game.bullets;
+    for (var index = 0; index < bullets.length; index += 1) {
+        if (bullets[index] === undefined) continue;
+        if (bullets[index].alive == false) {
+            bullets.splice(index, 1);
+            continue;
+        }
+
+        bullets[index].position.add(bullets[index].velocity);
+    }
+
     keyboardAction();
-    moveObjects();
+    if (player.canShoot > 0) player.canShoot -= 1;
 
-    renderer.render(screen.scene, screen.camera);
-    if (gameState) {
-        // requestAnimationFrame(animate);
-    }
+    // position the gun in front of the camera
+    renderer.render(game.scene, game.player.camera);
+    requestAnimationFrame(render);
 }
 
-/**keyboard action */
 function keyboardAction() {
-    if (keyboard[87]) {
-        // W key
-    } else if (keyboard[83]) {
-        // S key
-    } else if (keyboard[65]) {
-        // A key
-    } else if (keyboard[68]) {
-        // D key
-    }
+    let player = window.game.player;
+    let keyboard = window.game.keyboard;
+    if (keyboard[87]) player.key_w();
+    if (keyboard[83]) player.key_s();
+    if (keyboard[65]) player.key_a();
+    if (keyboard[68]) player.key_d();
+    if (keyboard[37]) player.key_lr(true);
+    if (keyboard[39]) player.key_lr(false);
+
     // shoot a bullet
     if (keyboard[32] && player.canShoot <= 0) {
-        //TODO: generate bullets
+        // spacebar key
+        // creates a bullet as a Mesh object
+        var bullet = new THREE.Mesh(new THREE.SphereGeometry(0.05, 8, 8), new THREE.MeshBasicMaterial({ color: 0xffffff }));
+        // this is silly.
+        // var bullet = models.pirateship.mesh.clone();
+
+        // position the bullet to come from the player's weapon
+        let camera = window.game.camera;
+        bullet.position.set(camera.position.x, camera.position.y + 0.15, camera.position.z);
+
+        // set the velocity of the bullet
+        bullet.velocity = new THREE.Vector3(-Math.sin(camera.rotation.y), 0, Math.cos(camera.rotation.y));
+
+        // after 1000ms, set alive to false and remove from scene
+        // setting alive to false flags our update code to remove
+        // the bullet from the bullets array
+        bullet.alive = true;
+        let scene = window.game.scene;
+        setTimeout(function () {
+            bullet.alive = false;
+            scene.remove(bullet);
+        }, 1000);
+
+        // add to scene, array, and set the delay to 10 frames
+        window.game.bullets.push(bullet);
+        scene.add(bullet);
+        player.canShoot = 10;
     }
 }
-
-/**
- * generate random new Objets
- */
-function generateObjects() {}
-
-/**
- * move Object points, collision process
- */
-function moveObjects() {
-    for (const meteo of meteos) meteo.move();
-    for (const planet of planets) planet.move();
-    for (const item of items) item.move();
-    for (const bullet of bullets) bullet.move();
-    //TODO: remove too far object
-    //TODO: collision check, work with player, Bullet
-}
-
-function keyDown(event) {
-    keyboard[event.keyCode] = true;
-}
-
-function keyUp(event) {
-    keyboard[event.keyCode] = false;
-}
-
+window.addEventListener("keydown", (event) => (window.game.keyboard[event.keyCode] = true));
+window.addEventListener("keyup", (event) => (window.game.keyboard[event.keyCode] = undefined));
 window.onload = init;
