+++ src/index.html
@@ -8,7 +8,7 @@
         <script type="text/javascript" src="./lib/OBJLoader.js"></script>
 
         <script type="text/javascript" src="./lib/GLTFLoader.js"></script>
-        <script type="module" src="./demo.js"></script>
+        <script type="module" src="./app.js"></script>
     </head>
     <body style="text-align: center">
         <span style="position: absolute; top: 0px; left: 0px"> WASD to move.<br />Arrow keys to turn. </span>
